fix(axios): Rewrite all client with Axios
Rewrite analytics client using the Axios
