+++ package.json
@@ -3,10 +3,10 @@
     "import": "node src/scripts/runTs.js import ts",
     "chart": "node src/scripts/runTs.js chart ts",
     "start": "node src/scripts/runTs.js start ts",
-    "dev": "node src/scripts/runTs.js dev ts",
+    "dev": "dotenv node src/scripts/runTs.js dev ts",
     "cluster": "node src/scripts/runTs.js cluster ts",
     "start:js": "node dist/src/scripts/runJs.js start js",
-    "dev:js": "node dist/src/scripts/runJs.js dev js",
+    "dev:js": "dotenv node dist/src/scripts/runJs.js dev js",
     "cluster:js": "node dist/src/scripts/runJs.js cluster js",
     "format": "prettier --write",
     "lint": "eslint",
@@ -64,6 +64,7 @@
     "concurrently": "7.0.0",
     "eslint": "8.10.0",
     "eslint-config-prettier": "8.4.0",
+    "dotenv-cli": "^6.0.0",
     "husky": "8.0.1",
     "jest": "26.6.3",
     "jest-sonar": "0.2.12",
