+++ src/analytics/api/index.ts
@@ -1,47 +1,40 @@
-import axios, { AxiosInstance } from "axios";
+import axios from "axios";
+import { Logger } from "../../logger";
 import { AnalyticsDataDto } from "./types";
+import { analyticsId } from "../../env";
 
-export class GoogleAnalyticsApi {
-  public static readonly url = "https://www.google-analytics.com";
-  public static readonly timeout = 60_000;
+const logger = new Logger("analytics");
 
-  private static readonly path = "collect";
+export const collectEvents = (data: AnalyticsDataDto): Promise<void> => {
+  if (!analyticsId) {
+    logger.error(
+      "Analytics Token is not provided!",
+      new Error("Analytics Token is not provided")
+    );
+    return Promise.resolve();
+  }
 
-  private readonly client: AxiosInstance;
+  const timeout = 10_000;
+  const logged = { ul: data.ul, cid: data.cid, uid: data.uid, t: data.t };
 
-  constructor() {
-    this.client = axios.create({
+  const url = new URL("https://www.google-analytics.com/collect");
+  Object.keys(data).forEach((key) => {
+    url.searchParams.set(key, data[key]);
+  });
+
+  return axios
+    .request({
       method: "GET",
-      baseURL: GoogleAnalyticsApi.url,
+      url: url.toString(),
       headers: {
         Accept: "application/json",
         "User-Agent":
           "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.105 Safari/537.36",
         "Content-Type": "application/json",
       },
-      timeout: GoogleAnalyticsApi.timeout,
+      timeout: timeout,
       responseType: "json",
-    });
-  }
-
-  public collect(dto: AnalyticsDataDto): Promise<void> {
-    const query = Object.keys(dto)
-      .map((key) => `${key}=${encodeURIComponent(dto[key])}`)
-      .join("&");
-    return this.request<void, void>(`${GoogleAnalyticsApi.path}?${query}`);
-  }
-
-  private request<Response, Data>(
-    methodName: string,
-    data?: Data
-  ): Promise<Response> {
-    const url = this.getApiUrl(methodName);
-    return this.client
-      .request<Response>({ url, data })
-      .then((response) => response.data);
-  }
-
-  private getApiUrl(methodName: string): string {
-    return `/${methodName}`;
-  }
-}
+    })
+    .then(() => logger.info("Analytic data has been collected", logged))
+    .catch((err) => logger.warn("Failed to collect analytic data", err));
+};
